requirements.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.2
fire==0.3.1
importlib-metadata==1.6.0
kombu==4.6.8
PyMySQL==0.9.3
pytz==2020.1
redis==3.5.2
six==1.14.0
termcolor==1.1.0
vine==1.3.0
websocket-client-py3==0.15.0
zipp==3.1.0
